refactor: BaseballGame 클래스를 별도 파일로 분리
